Move HA integration in proper folder and update the README file
